collapsed appbar modifications, breathing update
