+++ lib/core/debug_overlay.dart
@@ -2,11 +2,9 @@ import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:self_help/core/constants/routes_constants.dart';
-import 'package:self_help/pages/global_providers/app_overlay_provider.dart';
 import 'package:self_help/pages/global_providers/page_flow_provider.dart';
 import 'package:self_help/pages/global_providers/router_provider.dart';
 import 'package:self_help/pages/global_providers/user_auth_provider.dart';
-import 'package:self_help/services/services.dart';
 
 class DebugOverlay extends HookConsumerWidget {
   const DebugOverlay({super.key});
