+++ lib/pages/sos_landing/sos_landing.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:self_help/core/constants/flow_route_constant.dart';
 import 'package:self_help/core/constants/routes_constants.dart';
@@ -14,22 +15,35 @@ class SosLanding extends HookConsumerWidget {
   Widget build(BuildContext context, WidgetRef ref) {
     final localizations = AppLocalizations.of(context)!;
 
-    final routerListener = ref.watch(routerListenerProvider);
-    if (routerListener == RoutePaths.sosLanding) {
-      WidgetsBinding.instance.addPostFrameCallback(
-        (_) => ref.read(animatedAppBarProvider.notifier).updateState(
-              appBarType: AppBarType.fullScreen,
-              fullScreenTitle: localizations.sosMainTitle,
-              subtitle: localizations.sosMainInfo,
-              startCallback: () {
-                final provider = ref.read(pageFlowProvider.notifier);
-                provider.startFlow(FlowType.sos);
-              },
-              hasBackButton: true,
-            ),
-      );
-    }
+    ref.listen(
+      routerListenerProvider,
+      (previous, next) {
+        if (next != previous && next == RoutePaths.home) {
+          updateAppBar(ref, localizations);
+        }
+      },
+    );
+
+    useEffect(() {
+      updateAppBar(ref, localizations);
+      return null;
+    }, const []);
 
     return SizedBox.shrink();
+  }
+
+  void updateAppBar(WidgetRef ref, AppLocalizations localizations) {
+    WidgetsBinding.instance.addPostFrameCallback(
+      (_) => ref.read(animatedAppBarProvider.notifier).updateState(
+            appBarType: AppBarType.fullScreen,
+            fullScreenTitle: localizations.sosMainTitle,
+            subtitle: localizations.sosMainInfo,
+            startCallback: () {
+              final provider = ref.read(pageFlowProvider.notifier);
+              provider.startFlow(FlowType.sos);
+            },
+            hasBackButton: true,
+          ),
+    );
   }
 }
